Allow mcuboot tool to use cbor2 additional to cbor
The mcuboot imgtool uses the python module `cbor`.

An equivalent and updated package `cbor2` requires Python 3.7 or newer.

The arch packages provide a package `python-cbor2`, but no package for
`cbor`.

This patch makes it possible to use the system package by adding
support for the `cbor2` package additionally to the `cbor` package.